+++ puput/__init__.py
@@ -1,11 +1,10 @@
 __author__ = "Marc Tudurí"
 __email__ = "marctc@gmail.com"
-__version__ = "1.2.1"
+__version__ = "2.2.0"
 
 PUPUT_APPS = (
     # Wagtail apps
     "wagtail.contrib.legacy.richtext",
-    "wagtail.core",
     "wagtail.admin",
     "wagtail.documents",
     "wagtail.snippets",
@@ -16,12 +15,13 @@ PUPUT_APPS = (
     "wagtail.sites",
     "wagtail.contrib.redirects",
     "wagtail.contrib.forms",
+    "wagtail.contrib.search_promotions",
     "wagtail.contrib.sitemaps",
     "wagtail.contrib.routable_page",
+    "wagtail",
     # Third-party apps
     "taggit",
     "modelcluster",
-    "django_social_share",
     # Puput apps
     "puput",
     "wagtailmarkdown",